multi-value query parameters (#63)
* before multi-value query params

* multi-value query parameters

+++ services/mgmt/batch/src/package_2020_03/operations.rs
@@ -842,7 +842,7 @@ pub mod application_package {
         resource_group_name: &str,
         account_name: &str,
         application_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ListApplicationPackagesResult, list::Error> {
         let client = &operation_config.client;
@@ -1168,7 +1168,7 @@ pub mod application {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         account_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ListApplicationsResult, list::Error> {
         let client = &operation_config.client;
@@ -1424,7 +1424,7 @@ pub mod certificate {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         account_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         select: Option<&str>,
         filter: Option<&str>,
         subscription_id: &str,
@@ -1834,7 +1834,7 @@ pub mod private_link_resource {
         subscription_id: &str,
         resource_group_name: &str,
         account_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
     ) -> std::result::Result<ListPrivateLinkResourcesResult, list_by_batch_account::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -1967,7 +1967,7 @@ pub mod private_endpoint_connection {
         subscription_id: &str,
         resource_group_name: &str,
         account_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
     ) -> std::result::Result<ListPrivateEndpointConnectionsResult, list_by_batch_account::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -2167,7 +2167,7 @@ pub mod pool {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         account_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         select: Option<&str>,
         filter: Option<&str>,
         subscription_id: &str,
