+++ services/mgmt/batchai/src/package_2018_03/operations.rs
@@ -426,7 +426,7 @@ pub mod clusters {
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
         select: Option<&str>,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ClusterListResult, list::Error> {
         let client = &operation_config.client;
@@ -498,7 +498,7 @@ pub mod clusters {
         resource_group_name: &str,
         filter: Option<&str>,
         select: Option<&str>,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ClusterListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -899,7 +899,7 @@ pub mod jobs {
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
         select: Option<&str>,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<JobListResult, list::Error> {
         let client = &operation_config.client;
@@ -971,7 +971,7 @@ pub mod jobs {
         resource_group_name: &str,
         filter: Option<&str>,
         select: Option<&str>,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<JobListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -1044,8 +1044,8 @@ pub mod jobs {
         job_name: &str,
         outputdirectoryid: &str,
         directory: Option<&str>,
-        linkexpiryinminutes: Option<i64>,
-        maxresults: Option<i64>,
+        linkexpiryinminutes: Option<i32>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<FileListResult, list_output_files::Error> {
         let client = &operation_config.client;
@@ -1317,7 +1317,7 @@ pub mod file_servers {
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
         select: Option<&str>,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<FileServerListResult, list::Error> {
         let client = &operation_config.client;
@@ -1389,7 +1389,7 @@ pub mod file_servers {
         resource_group_name: &str,
         filter: Option<&str>,
         select: Option<&str>,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<FileServerListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
