+++ services/mgmt/keyvault/src/package_2018_02_14_preview/operations.rs
@@ -267,7 +267,7 @@ pub mod vaults {
     pub async fn list_by_resource_group(
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<VaultListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -314,7 +314,7 @@ pub mod vaults {
     }
     pub async fn list_by_subscription(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<VaultListResult, list_by_subscription::Error> {
         let client = &operation_config.client;
@@ -493,7 +493,7 @@ pub mod vaults {
     pub async fn list(
         operation_config: &crate::OperationConfig,
         filter: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list::Error> {
         let client = &operation_config.client;
