+++ services/mgmt/resources/src/package_resources_2016_09/operations.rs
@@ -338,7 +338,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentListResult, list::Error> {
         let client = &operation_config.client;
@@ -539,7 +539,7 @@ pub mod providers {
     }
     pub async fn list(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         expand: Option<&str>,
         subscription_id: &str,
     ) -> std::result::Result<ProviderListResult, list::Error> {
@@ -641,7 +641,7 @@ pub mod resource_groups {
         resource_group_name: &str,
         filter: Option<&str>,
         expand: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list_resources::Error> {
         let client = &operation_config.client;
@@ -969,7 +969,7 @@ pub mod resource_groups {
     pub async fn list(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceGroupListResult, list::Error> {
         let client = &operation_config.client;
@@ -1069,7 +1069,7 @@ pub mod resources {
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
         expand: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list::Error> {
         let client = &operation_config.client;
@@ -1922,7 +1922,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentOperationsListResult, list::Error> {
         let client = &operation_config.client;
