+++ services/mgmt/resources/src/package_resources_2019_05/operations.rs
@@ -422,7 +422,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         group_id: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentListResult, list_at_management_group_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -840,7 +840,7 @@ pub mod deployments {
     pub async fn list_at_subscription_scope(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentListResult, list_at_subscription_scope::Error> {
         let client = &operation_config.client;
@@ -1217,7 +1217,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -1419,7 +1419,7 @@ pub mod providers {
     }
     pub async fn list(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         expand: Option<&str>,
         subscription_id: &str,
     ) -> std::result::Result<ProviderListResult, list::Error> {
@@ -1521,7 +1521,7 @@ pub mod resources {
         resource_group_name: &str,
         filter: Option<&str>,
         expand: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -1672,7 +1672,7 @@ pub mod resources {
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
         expand: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list::Error> {
         let client = &operation_config.client;
@@ -2512,7 +2512,7 @@ pub mod resource_groups {
     pub async fn list(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceGroupListResult, list::Error> {
         let client = &operation_config.client;
@@ -2855,7 +2855,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         group_id: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_management_group_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -2950,7 +2950,7 @@ pub mod deployment_operations {
     pub async fn list_at_subscription_scope(
         operation_config: &crate::OperationConfig,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_subscription_scope::Error> {
         let client = &operation_config.client;
@@ -3044,7 +3044,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentOperationsListResult, list::Error> {
         let client = &operation_config.client;
