+++ services/mgmt/resources/src/package_resources_2019_07/operations.rs
@@ -514,7 +514,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         scope: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentListResult, list_at_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -1029,7 +1029,7 @@ pub mod deployments {
     pub async fn list_at_tenant_scope(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentListResult, list_at_tenant_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!("{}/providers/Microsoft.Resources/deployments/", &operation_config.base_path,);
@@ -1590,7 +1590,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         group_id: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentListResult, list_at_management_group_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -2226,7 +2226,7 @@ pub mod deployments {
     pub async fn list_at_subscription_scope(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentListResult, list_at_subscription_scope::Error> {
         let client = &operation_config.client;
@@ -2810,7 +2810,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -3063,7 +3063,7 @@ pub mod providers {
     }
     pub async fn list(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         expand: Option<&str>,
         subscription_id: &str,
     ) -> std::result::Result<ProviderListResult, list::Error> {
@@ -3127,7 +3127,7 @@ pub mod providers {
     }
     pub async fn list_at_tenant_scope(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         expand: Option<&str>,
     ) -> std::result::Result<ProviderListResult, list_at_tenant_scope::Error> {
         let client = &operation_config.client;
@@ -3323,7 +3323,7 @@ pub mod resources {
         resource_group_name: &str,
         filter: Option<&str>,
         expand: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -3523,7 +3523,7 @@ pub mod resources {
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
         expand: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list::Error> {
         let client = &operation_config.client;
@@ -4652,7 +4652,7 @@ pub mod resource_groups {
     pub async fn list(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceGroupListResult, list::Error> {
         let client = &operation_config.client;
@@ -5110,7 +5110,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         scope: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -5235,7 +5235,7 @@ pub mod deployment_operations {
     pub async fn list_at_tenant_scope(
         operation_config: &crate::OperationConfig,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_tenant_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -5366,7 +5366,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         group_id: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_management_group_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -5496,7 +5496,7 @@ pub mod deployment_operations {
     pub async fn list_at_subscription_scope(
         operation_config: &crate::OperationConfig,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_subscription_scope::Error> {
         let client = &operation_config.client;
@@ -5624,7 +5624,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentOperationsListResult, list::Error> {
         let client = &operation_config.client;
