+++ services/mgmt/resources/src/package_resources_2019_10/operations.rs
@@ -520,7 +520,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         scope: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentListResult, list_at_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -1109,7 +1109,7 @@ pub mod deployments {
     pub async fn list_at_tenant_scope(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentListResult, list_at_tenant_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!("{}/providers/Microsoft.Resources/deployments/", &operation_config.base_path,);
@@ -1749,7 +1749,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         group_id: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentListResult, list_at_management_group_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -2391,7 +2391,7 @@ pub mod deployments {
     pub async fn list_at_subscription_scope(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentListResult, list_at_subscription_scope::Error> {
         let client = &operation_config.client;
@@ -2981,7 +2981,7 @@ pub mod deployments {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -3234,7 +3234,7 @@ pub mod providers {
     }
     pub async fn list(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         expand: Option<&str>,
         subscription_id: &str,
     ) -> std::result::Result<ProviderListResult, list::Error> {
@@ -3298,7 +3298,7 @@ pub mod providers {
     }
     pub async fn list_at_tenant_scope(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         expand: Option<&str>,
     ) -> std::result::Result<ProviderListResult, list_at_tenant_scope::Error> {
         let client = &operation_config.client;
@@ -3494,7 +3494,7 @@ pub mod resources {
         resource_group_name: &str,
         filter: Option<&str>,
         expand: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -3694,7 +3694,7 @@ pub mod resources {
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
         expand: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list::Error> {
         let client = &operation_config.client;
@@ -4829,7 +4829,7 @@ pub mod resource_groups {
     pub async fn list(
         operation_config: &crate::OperationConfig,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceGroupListResult, list::Error> {
         let client = &operation_config.client;
@@ -5524,7 +5524,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         scope: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -5649,7 +5649,7 @@ pub mod deployment_operations {
     pub async fn list_at_tenant_scope(
         operation_config: &crate::OperationConfig,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_tenant_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -5780,7 +5780,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         group_id: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_management_group_scope::Error> {
         let client = &operation_config.client;
         let uri_str = &format!(
@@ -5910,7 +5910,7 @@ pub mod deployment_operations {
     pub async fn list_at_subscription_scope(
         operation_config: &crate::OperationConfig,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentOperationsListResult, list_at_subscription_scope::Error> {
         let client = &operation_config.client;
@@ -6038,7 +6038,7 @@ pub mod deployment_operations {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         deployment_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<DeploymentOperationsListResult, list::Error> {
         let client = &operation_config.client;
